* xml-stats moved from clixon-config.yang as state data to an rpc `datastats`in clixon-lib.yang * Two new plugin callbacks added   * ca_daemon: Called just after a server has "daemonized", ie put in background.   * ca_trans_commit_done: Called when all plugin commits have been done.

+++ lib/clixon/clixon_xml.h
@@ -154,7 +154,7 @@ typedef struct clixon_xml_vec clixon_xvec; /* struct defined in clicon_xml_vec.c
  */
 char     *xml_type2str(enum cxobj_type type);
 int       xml_stats_global(uint64_t *nr);
-size_t    xml_stats(cxobj *xt, uint64_t *nrp, size_t *szp);
+int       xml_stats(cxobj *xt, uint64_t *nrp, size_t *szp);
 char     *xml_name(cxobj *xn);
 int       xml_name_set(cxobj *xn, char *name);
 char     *xml_prefix(cxobj *xn);
